backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pandas
numpy
scikit-learn
joblib
python-multipart
requests
beautifulsoup4
googlesearch-python
google-genai
google-genai
python-dotenv
groq
httpx
